Sdtinstructions

See merge request lab2223_summer/armv8_25!8

+++ src/sdtInstruction.h
@@ -1,5 +1,6 @@
 #ifndef SDT_INSTRUCTION_H
 #define SDT_INSTRUCTION_H
+#include "immediateInstruction.h"
 
-void executeSdtInstruction(ComputerState*, const int32_t instruction);
+void executeSdtInstruction(ComputerState*, const int64_t instruction);
 #endif // SDT_INSTRUCTION_H
